Feat: findAllByWriterId의 writerId를 String 타입으로 변경

+++ src/main/kotlin/com/youquiz/quiz/handler/QuizHandler.kt
@@ -19,7 +19,7 @@ class QuizHandler(
 
     suspend fun findAllByWriterId(request: ServerRequest): ServerResponse =
         request.pathVariable("id").let {
-            ServerResponse.ok().bodyAndAwait(quizService.findAllByWriterId(it.toLong()))
+            ServerResponse.ok().bodyAndAwait(quizService.findAllByWriterId(it))
         }
 
     suspend fun findAllMarkedQuiz(request: ServerRequest): ServerResponse =
